Unified test code between platforms
