+++ platforms/opencl/tests/TestOpenCLCustomNonbondedForce.cpp
@@ -32,60 +32,6 @@
 #include "OpenCLTests.h"
 #include "TestCustomNonbondedForce.h"
 
-void testParallelComputation() {
-    System system;
-    const int numParticles = 200;
-    for (int i = 0; i < numParticles; i++)
-        system.addParticle(1.0);
-    CustomNonbondedForce* force = new CustomNonbondedForce("4*eps*((sigma/r)^12-(sigma/r)^6); sigma=0.5*(sigma1+sigma2); eps=sqrt(eps1*eps2)");
-    force->addPerParticleParameter("sigma");
-    force->addPerParticleParameter("eps");
-    for (int i = 0; i < numParticles; i++)
-        force->addParticle({0.5, 1.0});
-    system.addForce(force);
-    OpenMM_SFMT::SFMT sfmt;
-    init_gen_rand(0, sfmt);
-    vector<Vec3> positions(numParticles);
-    for (int i = 0; i < numParticles; i++)
-        positions[i] = Vec3(5*genrand_real2(sfmt), 5*genrand_real2(sfmt), 5*genrand_real2(sfmt));
-    for (int i = 0; i < numParticles; ++i)
-        for (int j = 0; j < i; ++j) {
-            Vec3 delta = positions[i]-positions[j];
-            if (delta.dot(delta) < 0.1)
-                force->addExclusion(i, j);
-        }
-    VerletIntegrator integrator1(0.01);
-    Context context1(system, integrator1, platform);
-    context1.setPositions(positions);
-    State state1 = context1.getState(State::Forces | State::Energy);
-    VerletIntegrator integrator2(0.01);
-
-    map<string, string> props;
-    string deviceIndex = platform.getPropertyValue(context1, OpenCLPlatform::OpenCLDeviceIndex());
-    props[OpenCLPlatform::OpenCLDeviceIndex()] = deviceIndex+","+deviceIndex;
-    string platformIndex = platform.getPropertyValue(context1, OpenCLPlatform::OpenCLPlatformIndex());
-    props[OpenCLPlatform::OpenCLPlatformIndex()] = platformIndex;
-
-    Context context2(system, integrator2, platform, props);
-    context2.setPositions(positions);
-    State state2 = context2.getState(State::Forces | State::Energy);
-    ASSERT_EQUAL_TOL(state1.getPotentialEnergy(), state2.getPotentialEnergy(), 1e-5);
-    for (int i = 0; i < numParticles; i++)
-        ASSERT_EQUAL_VEC(state1.getForces()[i], state2.getForces()[i], 1e-5);
-
-    // Try updating some parameters and see if they still match.
-
-    vector<double> params;
-    for (int i = 95; i < 102; i++)
-        force->setParticleParameters(i, {0.6, 2.0});
-    force->updateParametersInContext(context1);
-    force->updateParametersInContext(context2);
-    State state3 = context1.getState(State::Energy);
-    State state4 = context2.getState(State::Energy);
-    ASSERT_EQUAL_TOL(state3.getPotentialEnergy(), state4.getPotentialEnergy(), 1e-5);
-    ASSERT(fabs(state1.getPotentialEnergy()-state3.getPotentialEnergy()) > 0.1);
-}
-
 void runPlatformTests() {
     testParallelComputation();
 }
