+++ platforms/opencl/tests/TestOpenCLPeriodicTorsionForce.cpp
@@ -32,53 +32,6 @@
 #include "OpenCLTests.h"
 #include "TestPeriodicTorsionForce.h"
 
-void testParallelComputation() {
-    System system;
-    const int numParticles = 200;
-    for (int i = 0; i < numParticles; i++)
-        system.addParticle(1.0);
-    PeriodicTorsionForce* force = new PeriodicTorsionForce();
-    for (int i = 3; i < numParticles; i++)
-        force->addTorsion(i-3, i-2, i-1, i, 2, 1.1, i);
-    system.addForce(force);
-    vector<Vec3> positions(numParticles);
-    for (int i = 0; i < numParticles; i++)
-        positions[i] = Vec3(i, i%2, i%3);
-    VerletIntegrator integrator1(0.01);
-    Context context1(system, integrator1, platform);
-    context1.setPositions(positions);
-    State state1 = context1.getState(State::Forces | State::Energy);
-    VerletIntegrator integrator2(0.01);
-
-    map<string, string> props;
-    string deviceIndex = platform.getPropertyValue(context1, OpenCLPlatform::OpenCLDeviceIndex());
-    props[OpenCLPlatform::OpenCLDeviceIndex()] = deviceIndex+","+deviceIndex;
-    string platformIndex = platform.getPropertyValue(context1, OpenCLPlatform::OpenCLPlatformIndex());
-    props[OpenCLPlatform::OpenCLPlatformIndex()] = platformIndex;
-
-    Context context2(system, integrator2, platform, props);
-    context2.setPositions(positions);
-    State state2 = context2.getState(State::Forces | State::Energy);
-    ASSERT_EQUAL_TOL(state1.getPotentialEnergy(), state2.getPotentialEnergy(), 1e-5);
-    for (int i = 0; i < numParticles; i++)
-        ASSERT_EQUAL_VEC(state1.getForces()[i], state2.getForces()[i], 1e-5);
-
-    // Try updating some parameters and see if they still match.
-
-    for (int i = 95; i < 102; i++) {
-        int p1, p2, p3, p4, periodicity;
-        double phase, k;
-        force->getTorsionParameters(i, p1, p2, p3, p4, periodicity, phase, k);
-        force->setTorsionParameters(i, p1, p2, p3, p4, periodicity, phase+0.1, 2*k);
-    }
-    force->updateParametersInContext(context1);
-    force->updateParametersInContext(context2);
-    State state3 = context1.getState(State::Energy);
-    State state4 = context2.getState(State::Energy);
-    ASSERT_EQUAL_TOL(state3.getPotentialEnergy(), state4.getPotentialEnergy(), 1e-5);
-    ASSERT(fabs(state1.getPotentialEnergy()-state3.getPotentialEnergy()) > 0.1);
-}
-
 void runPlatformTests() {
     testParallelComputation();
 }
